Restructured data verification
Data Verification should be organized such that later features can be more easily implemented.
I commented out any backend stuff because I started laying out the logic for when Sam implements the user class returning reasons for failure.
