+++ Dont Eat Alone/Extensions/StringExtensions.swift
@@ -1,0 +1,20 @@
+//
+//  File.swift
+//  Dont Eat Alone
+//
+//  Created by Kameron Carr on 5/25/18.
+//  Copyright © 2018 Dont Eat Alone. All rights reserved.
+//
+
+import Foundation
+
+let __firstpart = "[A-Z0-9a-z]([A-Z0-9a-z._%+-]{0,30}[A-Z0-9a-z])?"
+let __serverpart = "([A-Z0-9a-z]([A-Z0-9a-z-]{0,30}[A-Z0-9a-z])?\\.){1,5}"
+let __emailRegex = __firstpart + "@" + __serverpart + "[A-Za-z]{2,6}"
+let __emailPredicate = NSPredicate(format: "SELF MATCHES %@", __emailRegex)
+
+extension String {
+    func isEmail() -> Bool {
+        return __emailPredicate.evaluate(with: self)
+    }
+}
